Add files via upload
The curve that we presented in the Notebook "Synchrotron Spectra Fitting " was not elegant due to the fact that it used a minimum amount of points. With that regard, we computated the full theoretical spectrum from scratch and used the fitted parameters in our models for better visualizations. In the first model, we used the hypothesis that the high energy observational data were coming from protons and next we reversed the hypothesis.